Move reminderFormSchema out of +page.server.ts
SvelteKit restricts exports from `+page.server.ts` to route-specific
symbols (`load`, `actions`, etc.). Exporting a `reminderFormSchema`
constant caused a 500 when visiting the edit page. Moved the schema
to `web-v4/src/routes/vehicles/[id]/reminders/_schemas.ts` (the
underscore prefix tells SvelteKit to ignore the file for routing)
and imported it from there.

When `add/reminder` gets migrated, it can import from the same
shared location.
